// src/slices/types.ts
export type SerializedBigNumber = string | null;
// ...
interface SerializedVaultUser {
  isLoading: boolean | null;
  userShares: SerializedBigNumber | null;
  cakeAtLastUserAction: SerializedBigNumber | null;
  lastDepositedTime: string | null;
  lastUserActionTime: string | null;
}

export interface SerializedLockedVaultUser extends SerializedVaultUser {
  lockStartTime: string | null;
  lockEndTime: string | null;
  userBoostedShare: SerializedBigNumber | null;
  locked: boolean | null;
  lockedAmount: SerializedBigNumber | null;
  currentPerformanceFee: SerializedBigNumber | null;
  currentOverdueFee: SerializedBigNumber | null;
}